Refactored ProjectileTag
