+++ game/src/ecs/components/skill/CurvedDamageProjectileSkill.java
@@ -3,7 +3,7 @@ package ecs.components.skill;
 import dslToGame.AnimationBuilder;
 import ecs.components.*;
 import ecs.components.collision.ICollide;
-import ecs.components.xp.ProjectileTag;
+import ecs.components.ProjectileTag;
 import ecs.damage.Damage;
 import ecs.entities.Entity;
 import graphic.Animation;
